Add ask things to Wolfram Alpha and Wikipedia
Now it's possible to ask things to them. Wolfram Alpha needs an API key.
Wikipedia doesn't.

From Wolfram Alpha, if it's a direct result it will
be spoken directly; if it's a more comprehensive result, the LLM will
be asked to summarize in sentences it first (since it doesn't come in
sentences).

From Wikipedia it's always directly spoken (my Raspberry Pi
isn't capable of summarizing a whole Wikipedia page in less than 40
years).