Auto-update LLM date and customizable nickname
Now the LLM knows the current date (not the time though), and eventually
it should be reset at the end of the day to update the date, but only
when the user is not using the LLM.

Also it's now possible to customize the name the LLM calls the user for.
Useful for females or any person who doesn't want VISOR calling them as
Sir. The rest of the project uses Sir though... That needs to be
changed too eventually.