User Settings changes (auto-get, for example)
Now User Settings are first loaded at startup on the client, and if they
can't be loaded, they're retrieved from the server, asking first for
the server domain and password. The Device Settings must already be
in place though. After this, the User Settings keep being saved forever
like the Gen Settings are. The file is no longer read from, also just
like the Gen Settings. If the file is loaded in the beginning and there
is network connection, it will only be used to get the server info to go
get the User Settings from the server anyway.

Aside from this, the Personal Constants are no longer mandatory. They
can all be empty - VISOR will just have limited functionality.

Also the app no longer uses the Registry to communicate when it should
show the window - it uses the communications channels.