Integration with Ollama API, chats added
It's still a work in progress but it's getting there. So far there's
integration with the Ollama API as opposed to using llama.cpp directly,
and I added the possibility of having chats, like on ChatGPT and others.

Because of that, the client UI has been updated to include a new Chats
tab on the Communicator screen.