Move libraries files out of now useless folders
They were used to differentiate the main and library packages when there
were testers for the libraries - now the only library that has a tester
is the ACD, because it's independent of the whole VISOR project. The
others need the client or the server running, so the tester is the
server or client and not a standalone app anymore.