Now possible to call models from other devices
Also possible to select 2 types of models: TEXT and VISION

And for the moment until the code is improved, this above and below is
disabled and only local server models will be used.

Method:
- Client sends text and model type request to server;
- Server asks itself and every client their local models;
- Clients return their available local models;
- Server sends the final request (with selected model included) to the
  client with the chosen model;
- Client executes the request on Ollama and keeps transmitting what it
  receives to the server;
- Server keeps receiving and redirecting the speeches to the client
  which initiated the request (which may or may not be the one with the
  chosen model).

Also to access Generated and User settings it's now necessary to specify
if it's to lock and unlock, only lock or only unlock the respective
mutex. Useful to prevent changes while using some setting - lock and in
the end of the usage, unlock the settings.