Func to report no models to Ollama from Android
For the moment I'm assuming Ollama or others can't be put on Android. In
case that's true, then this is a useful function tell the server there
are no models on the device (the device must report something always or
the server will keep waiting indefinitely). In case later I decide to
support that on Android, I won't use the function anymore, but it's here
for now.